refactor: UserInformation 컴포넌트 my 컴포넌트에서 공통 컴포넌트로 이동
